backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymongo==4.3.3
motor==3.1.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

pydantic==1.10.7

pandas==2.0.0
numpy==1.26.4
matplotlib==3.7.1
seaborn==0.12.2

openpyxl==3.1.2
fpdf2==2.7.4
jinja2==3.1.2

httpx==0.24.0
pytest==7.3.1

email-validator
python-dateutil
